requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
opencv-python>=4.5.0
numpy>=1.21.0
google-cloud-vision>=3.0.0
google-cloud-texttospeech>=2.14.0
google-generativeai>=0.3.0
Pillow>=8.0.0
requests>=2.25.0
gunicorn>=20.1.0
